Factor RoomPreviewCard out of SpaceRoomView
