+++ res/css/structures/_SpaceRoomView.scss
@@ -137,107 +137,6 @@ $SpaceRoomViewInnerWidth: 428px;
         }
     }
 
-    .mx_SpaceRoomView_preview {
-        padding: 32px 24px !important; // override default padding from above
-        margin: auto;
-        max-width: 480px;
-        box-sizing: border-box;
-        box-shadow: 2px 15px 30px $dialog-shadow-color;
-        border-radius: 8px;
-        position: relative;
-
-        // XXX remove this when spaces leaves Beta
-        .mx_BetaCard_betaPill {
-            position: absolute;
-            right: 24px;
-            top: 32px;
-        }
-
-        // XXX remove this when spaces leaves Beta
-        .mx_SpaceRoomView_preview_spaceBetaPrompt {
-            font-weight: $font-semi-bold;
-            font-size: $font-14px;
-            line-height: $font-24px;
-            color: $primary-content;
-            margin-top: 24px;
-            position: relative;
-            padding-left: 24px;
-
-            .mx_AccessibleButton_kind_link {
-                display: inline;
-                padding: 0;
-                font-size: inherit;
-                line-height: inherit;
-            }
-
-            &::before {
-                content: "";
-                position: absolute;
-                height: $font-24px;
-                width: 20px;
-                left: 0;
-                mask-repeat: no-repeat;
-                mask-position: center;
-                mask-size: contain;
-                mask-image: url('$(res)/img/element-icons/room/room-summary.svg');
-                background-color: $secondary-content;
-            }
-        }
-
-        .mx_SpaceRoomView_preview_inviter {
-            display: flex;
-            align-items: center;
-            margin-bottom: 20px;
-            font-size: $font-15px;
-
-            > div {
-                margin-left: 8px;
-
-                .mx_SpaceRoomView_preview_inviter_name {
-                    line-height: $font-18px;
-                }
-
-                .mx_SpaceRoomView_preview_inviter_mxid {
-                    line-height: $font-24px;
-                    color: $secondary-content;
-                }
-            }
-        }
-
-        > .mx_RoomAvatar_isSpaceRoom {
-            &.mx_BaseAvatar_image, .mx_BaseAvatar_image {
-                border-radius: 12px;
-            }
-        }
-
-        h1.mx_SpaceRoomView_preview_name {
-            margin: 20px 0 !important; // override default margin from above
-        }
-
-        .mx_SpaceRoomView_preview_topic {
-            font-size: $font-14px;
-            line-height: $font-22px;
-            color: $secondary-content;
-            margin: 20px 0;
-            max-height: 160px;
-            overflow-y: auto;
-        }
-
-        .mx_SpaceRoomView_preview_joinButtons {
-            margin-top: 20px;
-
-            .mx_AccessibleButton {
-                width: 200px;
-                box-sizing: border-box;
-                padding: 14px 0;
-
-                & + .mx_AccessibleButton {
-                    margin-left: 20px;
-                }
-            }
-        }
-    }
-
     .mx_SpaceRoomView_landing {
         display: flex;
         flex-direction: column;
