+++ res/css/views/elements/_FacePile.scss
@@ -15,6 +15,9 @@ limitations under the License.
 */
 
 .mx_FacePile {
+    display: flex;
+    align-items: center;
+
     .mx_FacePile_faces {
         display: inline-flex;
         flex-direction: row-reverse;
