+++ src/i18n/strings/en_EN.json
@@ -1872,6 +1872,9 @@
     "This room or space is not accessible at this time.": "This room or space is not accessible at this time.",
     "Try again later, or ask a room or space admin to check if you have access.": "Try again later, or ask a room or space admin to check if you have access.",
     "%(errcode)s was returned while trying to access the room or space. If you think you're seeing this message in error, please <issueLink>submit a bug report</issueLink>.": "%(errcode)s was returned while trying to access the room or space. If you think you're seeing this message in error, please <issueLink>submit a bug report</issueLink>.",
+    "Leave": "Leave",
+    "<inviter/> invites you": "<inviter/> invites you",
+    "To view %(roomName)s, you need an invite": "To view %(roomName)s, you need an invite",
     "Appearance": "Appearance",
     "Show rooms with unread messages first": "Show rooms with unread messages first",
     "Show previews of messages": "Show previews of messages",
@@ -1890,7 +1893,6 @@
     "Favourite": "Favourite",
     "Low Priority": "Low Priority",
     "Copy room link": "Copy room link",
-    "Leave": "Leave",
     "Video": "Video",
     "Connecting...": "Connecting...",
     "Connected": "Connected",
@@ -3115,8 +3117,6 @@
     "Results": "Results",
     "Rooms and spaces": "Rooms and spaces",
     "Search names and descriptions": "Search names and descriptions",
-    "<inviter/> invites you": "<inviter/> invites you",
-    "To view %(spaceName)s, you need an invite": "To view %(spaceName)s, you need an invite",
     "Welcome to <name/>": "Welcome to <name/>",
     "Random": "Random",
     "Support": "Support",
